add exitactual to sketchbase
